requirements.txt
django
pyyaml
watchdog
